Fixed bug and updated CalcServer and Infocenter

+++ Dev/AppUnits/Atdi.AppUnits.Sdrn.CalcServer/DataModel/ProjectMapData.cs
@@ -46,7 +46,7 @@ namespace Atdi.AppUnits.Sdrn.CalcServer.DataModel
 
 		public int OwnerStepsNumber => this.OwnerAxisXNumber * this.OwnerAxisYNumber;
 
-		public long RectArea => (this.LowerRightX - this.UpperLeftX) * (this.UpperLeftY - this.LowerRightY);
+		public ulong RectArea => ((ulong)this.LowerRightX - (ulong)this.UpperLeftX) * ((ulong)this.UpperLeftY - (ulong)this.LowerRightY);
 
 		public bool Has(int x, int y)
 		{
